Separating act.item.cpp has been finished.
