+++ src/dg_script/dg_scripts.cpp
@@ -59,7 +59,7 @@ extern IndexData *mob_index;
 extern TimeInfoData time_info;
 const char *GetSpellName(int num);
 
-extern int can_take_obj(CharData *ch, ObjData *obj);
+extern bool CanTakeObj(CharData *ch, ObjData *obj);
 extern void split_or_clan_tax(CharData *ch, long amount);
 
 // external functions
@@ -3123,7 +3123,7 @@ void find_replacement(void *go,
 			}
 			if ((*subfield == UID_CHAR) || (*subfield == UID_CHAR_ALL)) {
 				char_to = find_char(atoi(subfield + 1));
-				if (!(char_to && can_take_obj(char_to, o))) {
+				if (!(char_to && CanTakeObj(char_to, o))) {
 					trig_log(trig, "object.put: субъект-приемник не найден или не может нести этот объект");
 					return;
 				}
