+++ src/game_crafts/fry.h
@@ -1,0 +1,8 @@
+#ifndef BYLINS_SRC_GAME_CRAFTS_FRY_H_
+#define BYLINS_SRC_GAME_CRAFTS_FRY_H_
+
+class CharData;
+
+void do_fry(CharData *ch, char *argument, int/* cmd*/, int /*subcmd*/);
+
+#endif //BYLINS_SRC_GAME_CRAFTS_FRY_H_
