+++ src/utils/utils_char_obj.inl
@@ -1,10 +1,14 @@
 #ifndef __CHAR_OBJ_UTILS_HPP__
 #define __CHAR_OBJ_UTILS_HPP__
 
+#include "handler.h"
 #include "structs/structs.h"
 #include "entities/char_data.h"
 #include "entities/obj_data.h"
+#include "game_mechanics/named_stuff.h"
 #include "utils.h"
+
+extern int invalid_anti_class(CharData *ch, const ObjData *obj);
 
 inline bool INVIS_OK_OBJ(const CharData *sub, const ObjData *obj) {
 	return !obj->has_flag(EObjFlag::kInvisible)
@@ -50,6 +54,48 @@ inline bool CAN_GET_OBJ(const CharData *ch, const ObjData *obj) {
 		&& CAN_SEE_OBJ(ch, obj))
 		&& !(ch->IsNpc()
 			&& obj->has_flag(EObjFlag::kBloody));
+}
+
+inline bool CanTakeObj(CharData *ch, ObjData *obj) {
+	if (IS_CARRYING_N(ch) >= CAN_CARRY_N(ch)
+		&& GET_OBJ_TYPE(obj) != EObjType::kMoney) {
+		act("$p: Вы не могете нести столько вещей.", false, ch, obj, nullptr, kToChar);
+		return false;
+	} else if ((IS_CARRYING_W(ch) + GET_OBJ_WEIGHT(obj)) > CAN_CARRY_W(ch)
+		&& GET_OBJ_TYPE(obj) != EObjType::kMoney) {
+		act("$p: Вы не в состоянии нести еще и $S.", false, ch, obj, nullptr, kToChar);
+		return false;
+	} else if (!(CAN_WEAR(obj, EWearFlag::kTake))) {
+		act("$p: Вы не можете взять $S.", false, ch, obj, nullptr, kToChar);
+		return false;
+	} else if (invalid_anti_class(ch, obj)) {
+		act("$p: Эта вещь не предназначена для вас!", false, ch, obj, nullptr, kToChar);
+		return false;
+	} else if (NamedStuff::check_named(ch, obj, false)) {
+		if (!NamedStuff::wear_msg(ch, obj))
+			act("$p: Эта вещь не предназначена для вас!", false, ch, obj, nullptr, kToChar);
+		return false;
+	}
+	return true;
+}
+
+inline void weight_change_object(ObjData *obj, int weight) {
+	ObjData *tmp_obj;
+	CharData *tmp_ch;
+
+	if (obj->get_in_room() != kNowhere) {
+		obj->set_weight(std::max(1, GET_OBJ_WEIGHT(obj) + weight));
+	} else if ((tmp_ch = obj->get_carried_by())) {
+		ExtractObjFromChar(obj);
+		obj->set_weight(std::max(1, GET_OBJ_WEIGHT(obj) + weight));
+		PlaceObjToInventory(obj, tmp_ch);
+	} else if ((tmp_obj = obj->get_in_obj())) {
+		ExtractObjFromObj(obj);
+		obj->set_weight(std::max(1, GET_OBJ_WEIGHT(obj) + weight));
+		PlaceObjIntoObj(obj, tmp_obj);
+	} else {
+		log("SYSERR: Unknown attempt to subtract weight from an object.");
+	}
 }
 
 #endif // __CHAR_OBJ_UTILS_HPP__
